Runtime annotation (N#k) grammar updated and code refactoring
The runtime rule N#k was implemented as N%k. Now the grammar is updated and the use of this rule is according to GODA's article. The rule N#k is working properly and N%k no longer works.

RTParser and RTRegex now force antlr to throw an exception in the face of grammar error (instead of handling it).

Antlr generated files are in target/generated-sources/antlr4. The parses were updated to import theses classes.